+++ ci/casper_updater/src/main.rs
@@ -215,7 +215,7 @@ fn main() {
             &regex_data::smart_contracts_contract::DEPENDENT_FILES,
         ),
         Package::cargo(
-            "smart_contracts/v2/sdk_sys",
+            "smart_contracts/vm2/sdk_sys",
             &regex_data::smart_contracts_sdk_sys::DEPENDENT_FILES,
         ),
         Package::cargo(
@@ -223,7 +223,7 @@ fn main() {
             &regex_data::smart_contracts_sdk::DEPENDENT_FILES,
         ),
         Package::cargo(
-            "smart_contracts/v2/macros",
+            "smart_contracts/vm2/macros",
             &regex_data::smart_contracts_macros::DEPENDENT_FILES,
         ),
         Package::cargo(